requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.26.0
pandas==2.1.1
scikit-learn==1.3.1
imbalanced-learn==0.10.1
torch==2.1.0
transformers==4.34.0
owlready2==0.39
joblib==1.3.1
python-dotenv==1.0.0